app/requirements.txt
fastapi-slim==0.111.0
sqladmin
uvicorn[standard]==0.29.0
asyncpg
alembic==1.13.1
pydantic-settings==2.2.1